JQuery Link
Linked JQuery to HTML page to access it for use in my javascript page. Added an ordered list that was empty that will eventually be used for listing items in the to-do list. An input element was added so that the user has a visible box to input text into. Also added some javascript so far that will allow items to be added to list and if the user doesnt input any text, the app will alert them to type something