Created To-Do List
Used HTML and CSS to style a to-do list. Javascript was used to allow functionality for the list. The Javascript allows you to add items to the list, cross them out by double clicking on them, delete them by clicking on the X button, and alerts if nothing is typed into the input bar.